doc: format command help correctly with keeping doc examples (#169)
Example in the helpers are now unwrapped without modifying the documentation
