+++ copernicusmarine/core_functions/click_custom_class.py
@@ -1,28 +1,11 @@
-import functools
 import inspect
 import logging
-from typing import Any, Callable, Dict
 
 import click
 
+from copernicusmarine.core_functions.utils import log_deprecated_message
+
 logger = logging.getLogger("copernicusmarine")
-
-
-def get_deprecated_message(old_value, preferred_value):
-    return (
-        f"'{old_value}' has been deprecated, use '{preferred_value}' instead"
-    )
-
-
-def log_deprecated_message(old_value, preferred_value):
-    logger.warning(get_deprecated_message(old_value, preferred_value))
-
-
-def raise_both_old_and_new_value_error(old_value, new_value):
-    raise TypeError(
-        f"Received both {old_value} and {new_value} as arguments! "
-        f"{get_deprecated_message(old_value, new_value)}"
-    )
 
 
 class DeprecatedClickOption(click.Option):
@@ -32,7 +15,7 @@ class DeprecatedClickOption(click.Option):
         super().__init__(*args, **kwargs)
 
 
-class DeprecatedClickOptionsCommand(click.Command):
+class CustomClickOptionsCommand(click.Command):
     def make_parser(self, ctx):
         parser = super().make_parser(ctx)
 
@@ -68,25 +51,10 @@ class DeprecatedClickOptionsCommand(click.Command):
 
         return parser
 
-
-def deprecated_python_option(**aliases: str) -> Callable:
-    def deco(f: Callable):
-        @functools.wraps(f)
-        def wrapper(*args, **kwargs):
-            rename_kwargs(f.__name__, kwargs, aliases)
-            return f(*args, **kwargs)
-
-        return wrapper
-
-    return deco
-
-
-def rename_kwargs(
-    func_name: str, kwargs: Dict[str, Any], aliases: Dict[str, str]
-):
-    for alias, new in aliases.items():
-        if alias in kwargs:
-            if new in kwargs:
-                raise_both_old_and_new_value_error(alias, new)
-            log_deprecated_message(alias, new)
-            kwargs[new] = kwargs.pop(alias)
+    def format_epilog(self, ctx, formatter):
+        if self.epilog:
+            formatter.write_paragraph()
+            for line in self.epilog.split("\n"):
+                if ".. code-block::" in line:
+                    continue
+                formatter.write(line + "\n")
