+++ copernicusmarine/python_interface/get.py
@@ -1,12 +1,12 @@
 import pathlib
 from typing import Optional, Union
 
-from copernicusmarine.core_functions.deprecated import deprecated_python_option
 from copernicusmarine.core_functions.deprecated_options import (
     DEPRECATED_OPTIONS,
 )
 from copernicusmarine.core_functions.get import get_function
 from copernicusmarine.core_functions.models import ResponseGet
+from copernicusmarine.core_functions.utils import deprecated_python_option
 from copernicusmarine.python_interface.exception_handler import (
     log_exception_and_exit,
 )
