+++ copernicusmarine/python_interface/subset.py
@@ -2,7 +2,6 @@ import pathlib
 from datetime import datetime
 from typing import List, Optional, Union
 
-from copernicusmarine.core_functions.deprecated import deprecated_python_option
 from copernicusmarine.core_functions.deprecated_options import (
     DEPRECATED_OPTIONS,
 )
@@ -16,6 +15,7 @@ from copernicusmarine.core_functions.models import (
     VerticalDimensionOutput,
 )
 from copernicusmarine.core_functions.subset import subset_function
+from copernicusmarine.core_functions.utils import deprecated_python_option
 from copernicusmarine.python_interface.exception_handler import (
     log_exception_and_exit,
 )
